Refactor github git API into its own file
