+++ lib/platform/github/util.js
@@ -1,0 +1,11 @@
+module.exports = {
+  expandError,
+};
+
+function expandError(err) {
+  return {
+    err,
+    message: err.message,
+    body: err.response ? err.response.body : undefined,
+  };
+}
